fix: wrong type for filtering step in parquet files

+++ wayang-benchmark/src/main/java/org/apache/wayang/apps/parquet_csv/YelpBench.java
@@ -93,7 +93,7 @@ public class YelpBench {
         for (int i = 0; i < runs; i++) {
             CountDataQuantaBuilder<Record> parquet = planBuilder
                     .readParquet(fileSource)
-                    .filter(l -> l.getInt(0) == 0).withName("Remove non-zero labels")
+                    .filter(l -> l.getLong(0) == 0).withName("Remove non-zero labels")
                     .count();
 
             long startTime = System.currentTimeMillis();
